+++ src/flextable.rs
@@ -396,8 +396,8 @@ impl FlexTable {
             .map(|datatype| {
                 match datatype {
                     FlexDataType::Dbl => Cell::new("f64"),
-                    FlexDataType::Uint => Cell::new("u32"),
-                    FlexDataType::Int => Cell::new("i64"),
+                    FlexDataType::Uint => Cell::new("usize"),
+                    FlexDataType::Int => Cell::new("isize"),
                     FlexDataType::Char => Cell::new("char"),
                     FlexDataType::Str => Cell::new("str"),
                     FlexDataType::NA => Cell::new("n/a")
