+++ src/helper.rs
@@ -37,7 +37,7 @@ pub fn generate_flexdata_from_str(token: &str, datatype: &FlexDataType) -> FlexD
             }
         },
         FlexDataType::Int => {
-            if let Some( value ) = token.parse::<i64>().ok() {
+            if let Some( value ) = token.parse::<i32>().ok() {
                 FlexData::Int( value )
             } else {
                 FlexData::NA
@@ -91,7 +91,7 @@ pub fn convert(x: &FlexData, datatype: &FlexDataType) -> FlexData {
             match datatype {
                 FlexDataType::Str => FlexData::Str( format!("{}", val) ),
                 FlexDataType::Dbl => FlexData::Dbl( *val as f64 ),
-                FlexDataType::Int => FlexData::Int( *val as i64 ),
+                FlexDataType::Int => FlexData::Int( *val as i32 ),
                 FlexDataType::Uint => FlexData::Uint( *val as u32 ),
                 _ => FlexData::NA
             }
@@ -100,7 +100,7 @@ pub fn convert(x: &FlexData, datatype: &FlexDataType) -> FlexData {
             match datatype {
                 FlexDataType::Str => FlexData::Str( format!("{}", val) ),
                 FlexDataType::Dbl => FlexData::Dbl( *val as f64 ),
-                FlexDataType::Int => FlexData::Int( *val as i64 ),
+                FlexDataType::Int => FlexData::Int( *val as i32 ),
                 FlexDataType::Uint => FlexData::Uint( *val as u32 ),
                 _ => FlexData::NA
             }
@@ -109,7 +109,7 @@ pub fn convert(x: &FlexData, datatype: &FlexDataType) -> FlexData {
             match datatype {
                 FlexDataType::Str => FlexData::Str( format!("{}", val) ),
                 FlexDataType::Dbl => FlexData::Dbl( *val as f64 ),
-                FlexDataType::Int => FlexData::Int( *val as i64 ),
+                FlexDataType::Int => FlexData::Int( *val as i32 ),
                 FlexDataType::Uint => FlexData::Uint( *val as u32 ),
                 _ => FlexData::NA
             }
